+++ Data/SBV/Provers/Prover.hs
@@ -24,7 +24,8 @@ module Data.SBV.Provers.Prover (
        , allSat, allSatWith
        , isVacuous, isVacuousWith
        , solve
-       , SatModel(..), Modelable(..), displayModels, extractModels, getModelDictionaries, getModelValues
+       , SatModel(..), Modelable(..), displayModels, extractModels
+       , getModelDictionaries, getModelValues, getModelUninterpretedValues
        , boolector, cvc4, yices, z3, defaultSMTCfg
        , compileToSMTLib, generateSMTBenchmarks
        , sbvCheckSolverInstallation